requirements.txt
altair==5.4.0
archspec==0.2.1
asgiref==3.8.1
attrs==24.2.0
blinker==1.8.2
boltons==23.0.0
boto3==1.35.35
botocore==1.35.35
brotli==1.0.9
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
#conda==24.1.2
#conda-libmamba-solver==23.11.1
#conda-package-handling==2.2.0
#conda-package-streaming==0.9.0
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
distro==1.9.0
Django==5.1.1
django-filter==24.3
django-storages==1.14.4
fonttools==4.53.1
frozendict==2.4.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==23.0.0
h11==0.14.0
idna==3.6
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
#libmambapy==1.5.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
#menuinst==2.0.2
missingno==0.4.1
narwhals==1.5.0
numpy==2.1.0
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
parsel==1.9.1
patsy==0.5.6
pillow==10.4.0
platformdirs==3.11.0
plotly==5.23.0
plotly-express==0.4.0
pluggy==1.0.0
protobuf==5.27.3
pyarrow==17.0.0
pycosat==0.6.6
pycparser==2.21
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.17.21
s3transfer==0.10.2
scipy==1.14.1
seaborn==0.13.2
selectorlib==0.16.0
selenium==4.24.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
statsmodels==0.14.2
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.2
trio==0.26.2
trio-websocket==0.11.1
truststore==0.8.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
w3lib==2.2.1
webdriver-manager==4.0.2
websocket-client==1.8.0
whitenoise==6.7.0
wsproto==1.2.0
zstandard==0.22.0